feat: create entities for jpa repository
1. Create jpa entities
2. Create migration v1 for create tables in DB
3. Add dependencies for spring

+++ restaurant-application/src/main/java/com/vr61v/in/AdminService.java
@@ -23,7 +23,7 @@ public interface AdminService extends BasicUserService {
     Product createProduct(Product product);
     Product getProductById(UUID productId) throws NotFoundException;
     List<Product> getAllProducts();
-    Product updateProduct(UUID productId, UUID productCategoryId, Integer price, String name, Float weight, String composition, String description) throws NotFoundException;
+    Product updateProduct(UUID productId, UUID productCategoryId, Integer price, String name, Integer weight, String composition, String description) throws NotFoundException;
     UUID deleteProduct(UUID productId) throws NotFoundException;
 
     ProductCategory createProductCategory(ProductCategory productCategory);
